backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
resend==2.10.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
aiofiles==24.1.0
